Move the raw node GET onto the generic nodes route
